pack: 모니터링 관련 패키지 수정
