+++ src/main/java/org/jun/saemangeum/pipeline/application/alarm/AlarmBuilder.java
@@ -1,0 +1,8 @@
+package org.jun.saemangeum.pipeline.application.alarm;
+
+import org.jun.saemangeum.pipeline.domain.AlarmPayload;
+
+@FunctionalInterface
+public interface AlarmBuilder {
+    AlarmPayload build();
+}
