--- src/main/java/org/jun/saemangeum/pipeline/application/monitor/alarm/Alarm.java
@@ -1,5 +1,0 @@
-package org.jun.saemangeum.pipeline.application.monitor.alarm;
-
-public interface Alarm {
-    void sendAlarm(AlarmBuilder alarmBuilder, Object... args);
-}
