--- src/main/java/org/jun/saemangeum/pipeline/application/monitor/alarm/AlarmBuilder.java
@@ -1,6 +1,0 @@
-package org.jun.saemangeum.pipeline.application.monitor.alarm;
-
-@FunctionalInterface
-public interface AlarmBuilder {
-    AlarmPayload build();
-}
