+++ src/main/java/org/jun/saemangeum/pipeline/application/schedule/PipelineScheduler.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class DataProcessScheduler {
+public class PipelineScheduler {
 
     private final PipelineService pipelineService;
