+++ src/main/java/org/jun/saemangeum/pipeline/application/service/AlarmService.java
@@ -1,6 +1,12 @@
-package org.jun.saemangeum.pipeline.application.monitor.alarm;
+package org.jun.saemangeum.pipeline.application.service;
 
 import lombok.RequiredArgsConstructor;
+import org.jun.saemangeum.pipeline.application.alarm.AlarmProcess;
+import org.jun.saemangeum.pipeline.domain.Alarm;
+import org.jun.saemangeum.pipeline.domain.AlarmMessage;
+import org.jun.saemangeum.pipeline.domain.AlarmPayload;
+import org.jun.saemangeum.pipeline.domain.AlarmType;
+import org.jun.saemangeum.pipeline.presentation.alarm.*;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
