+++ src/main/java/org/jun/saemangeum/pipeline/application/service/PipelineService.java
@@ -4,7 +4,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.jun.saemangeum.global.domain.Content;
 import org.jun.saemangeum.global.service.ContentService;
 import org.jun.saemangeum.pipeline.application.collect.base.Refiner;
-import org.jun.saemangeum.pipeline.application.monitor.alarm.*;
 import org.jun.saemangeum.pipeline.infrastructure.dto.EmbeddingJob;
 import org.jun.saemangeum.pipeline.infrastructure.queue.EmbeddingWorkerService;
 import org.springframework.core.task.TaskExecutor;
