+++ src/main/java/org/jun/saemangeum/pipeline/domain/Alarm.java
@@ -1,0 +1,7 @@
+package org.jun.saemangeum.pipeline.domain;
+
+import org.jun.saemangeum.pipeline.application.alarm.AlarmBuilder;
+
+public interface Alarm {
+    void sendAlarm(AlarmBuilder alarmBuilder, Object... args);
+}
