--- src/main/java/org/jun/saemangeum/pipeline/presentation/AdminDTO.java
@@ -1,7 +1,0 @@
-package org.jun.saemangeum.pipeline.presentation;
-
-public record AdminDTO(
-        String username,
-        String password
-) {
-}
